% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.bootstrap.R
\name{cor.bootstrap}
\alias{cor.bootstrap}
\title{Calculate bootrstapped correlation coefficents for genes from matrix}
\usage{
cor.bootstrap(
  mtx,
  target,
  nbootstrap = nbootstrap,
  method = "pearson",
  nThreads = 10
)
}
\arguments{
\item{mtx}{matrix (transcripts x cells) containing scaled expression values}

\item{target}{Gene for wich correlation analyses against all other genes should be calculated}

\item{nbootstrap}{Number of bootrstrap iterations to perform}

\item{method}{Method to calculate correlation}

\item{nThreads}{number of Threads to perform computation on}
}
\value{
data.frame containing Tukey Box and Wishker Plot statistics (columns) for all transcripts (rows) in the matrix
}
\description{
Calculate bootrstapped correlation coefficents for genes from matrix
}
\details{
This function creates an ROC plot. Returns or prints a ggplot object.
}
