% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.plot.R
\name{cor.plot}
\alias{cor.plot}
\title{Calculate and visualize bootstrapped correlation coefficients for genes from a SEURAT object}
\usage{
cor.plot(
  corrtbl,
  GOI = NULL,
  bins = 400,
  alpha = 1,
  probs = c(0.75, 0.95, 0.99),
  font.size = 8
)
}
\arguments{
\item{corrtbl}{A correlation table as generated by the corr.transcripts function.}

\item{GOI}{Gene of Interest (GOI) to be highlighted in the plot. Default is NULL.}

\item{bins}{Number of bins to split data in for the histogram.}

\item{alpha}{Transparency of the histogram bars (0 for fully transparent, 1 for opaque).}

\item{probs}{Percentile borders to plot on the histogram (default is c(0.75, 0.95, 0.99)).}

\item{font.size}{Font size for axis labels and text annotations.}
}
\value{
A ggplot2 object containing the correlation coefficient histogram and, if specified, a box and whisker plot for the GOI.
}
\description{
This function generates a histogram of bootstrapped correlation coefficients for genes from a correlation table, along with an optional box and whisker plot for a Gene of Interest (GOI).
}
\seealso{
\code{\link{corr.transcripts}}
}
