% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.transcripts.R
\name{corr.transcripts}
\alias{corr.transcripts}
\title{Calculate bootrstapped correlation coefficents for genes from SEURAT object}
\usage{
corr.transcripts(
  object,
  target,
  ident.use = NULL,
  thresh = 0,
  method = "pearson",
  nbootstrap = 30,
  nThreads = 10,
  slot = "data"
)
}
\arguments{
\item{object}{Seurat object}

\item{target}{Gene for which correlation analyses against all other genes should be calculated}

\item{ident.use}{Only include particular clusters, as defined in Seurat object}

\item{thresh}{Only consider genes that are expressed in a percentage of cells greater than thresh}

\item{method}{Method to calculate correlation (default is "pearson")}

\item{nbootstrap}{Number of bootstrap iterations to perform (default is 30)}

\item{nThreads}{Number of threads to perform computation on (default is 10)}

\item{slot}{Slot to pull data from (default is "data")}
}
\value{
Data frame containing bootstrapped correlation coefficients
}
\description{
Calculate bootrstapped correlation coefficents for genes from SEURAT object
}
\details{
This function calculates bootstrapped correlation coefficients for a target gene against all other genes.
}
